To make bibliography compatible with Overleaf
The original template is not compatible with Overleaf (it gives errors and do not show the correct references in the bibliography section). The current changes make it compatible. The package "chapterbib" (with the option sectionbib for natbib) allows local references for each paper and the overview chapter, without the need to compile each bibliography separately. The line to input the bibliography after each section is no longer necessary (but kept as comment, at least in the near future). The new code was tested with Overleaf and pdflatex, for the original example (not tested with lualatex or any other PhD thesis). Disclaimer: all disclaimers of the original template applies to these changes (also, no Justin Biebers were harmed in the making of these changes).